refactor: Move data fetching into server side
