+++ app/projects/[projectId]/_component/MarkdownViewer.tsx
@@ -1,0 +1,14 @@
+import Markdown from 'react-markdown'
+import { Project } from '@/app/_types/project'
+
+export const MarkdownViewer = ({
+  project,
+}: {
+  project: Project,
+}) => {
+  return (
+    <Markdown className="prose prose-sm prose-invert">
+      {project.description}
+    </Markdown>
+  )
+}
