+++ app/projects/[projectId]/_component/MicroNDA.tsx
@@ -1,3 +1,4 @@
+'use client'
 import { addRole } from '../action'
 
 const MicroNDA = ({
