+++ app/projects/[projectId]/loader.ts
@@ -80,52 +80,3 @@ WHERE
     return []
   }
 }
-
-const  getFile = async (projectId: string, fileName: string) => {
-  const projectUri = projectsKey + projectId + '/';
-  const listParams = {
-    Bucket: transbucket,
-    Prefix: projectUri,
-    Delimiter: '/'
-  };
-  const listVersion = await client.send(new ListObjectsV2Command(listParams));
-  const folders = listVersion.CommonPrefixes?.map(prefix => prefix.Prefix);
-
-  const latestVersion = folders?.map(folder => folder?.split('/')[2])
-    .filter(folder => !isNaN(Number(folder)))
-    .sort((a, b) => Number(b) - Number(a))[0];
-
-  const fileKey = projectUri + latestVersion + '/' + fileName;
-  const data = await client.send(new ListObjectsV2Command({ Bucket: transbucket, Prefix: fileKey }));
-
-  if (data.Contents?.length && data.Contents.length > 0) {
-    const href = `https://${transbucket}.s3.${bucketRegion}.amazonaws.com/${fileKey}`;
-    return Response.json({ url: href })
-  }
-}
-
-export const downloadLoader = async (projectId: string) => {
-  const projectUri = projectsKey + projectId;
-  const filesData = await client.send(new ListObjectsV2Command({
-    Bucket: bucketName,
-    Prefix: projectUri,
-  }))
-
-  const filePromises = filesData.Contents?.map(async file => {
-    const fileData = await client.send(new GetObjectCommand({
-      Bucket: bucketName,
-      Key: file.Key
-    }))
-
-    if (fileData.Body) {
-      const byteArray = await fileData.Body.transformToByteArray()  // Uint8Array
-      const blob = new Blob([byteArray], { type: fileData.ContentType })  // blob
-      return {
-        key: file.Key,
-        blob: blob,
-      }
-    }
-  })
-
-  return await Promise.all(filePromises || [])
-}
